requirements.txt
transformers==4.12.5
sentence-transformers==2.2.0
umap-learn==0.5.3